fixed some type conversion conflicts between floats and doubles in Trip.java and related files

+++ extensions/health/src/main/java/de/tum/bgu/msm/health/data/Trip.java
@@ -43,8 +43,8 @@ public class Trip implements Id {
     private double matsimTravelDistance = 0.;
 
     private double marginalMetHours = 0.;
-    private Map<String, Float> travelRiskMap = new HashMap<>();
-    private Map<String, Float> travelExposureMap = new HashMap<>();
+    private Map<String, Double> travelRiskMap = new HashMap<>();
+    private Map<String, Double> travelExposureMap = new HashMap<>();
     private double travelNoiseExposure = 0.;
     private double travelNdviExposure = 0.;
 
@@ -147,14 +147,14 @@ public class Trip implements Id {
 
     public Map<String, Double> getTravelRiskMap() { return travelRiskMap; }
 
-    public void updateTravelRiskMap(Map<String, Float> newRisks) {
+    public void updateTravelRiskMap(Map<String, Double> newRisks) {
         // newRisks.forEach((k, v) -> travelRiskMap.merge(k, v, (v1, v2) -> v1 + v2 - v1*v2));
         newRisks.forEach((k, v) -> travelRiskMap.merge(k, v, (v1, v2) -> v1 + v2));
     }
 
-    public Map<String, Float> getTravelExposureMap() { return travelExposureMap; }
+    public Map<String, Double> getTravelExposureMap() { return travelExposureMap; }
 
-    public void updateTravelExposureMap(Map<String, Float> newExposures) {
+    public void updateTravelExposureMap(Map<String, Double> newExposures) {
         newExposures.forEach((k, v) -> travelExposureMap.merge(k, v, (v1, v2) -> v1 + v2));
     }
